Fire Employee bugfix
The patchers wanted to reduce the level from fired sim, because of the
exploit, firing and marrying a high level sim.
But the "reduce level" does not work.

So I build a blockade instead. A fired sim can't be courtlover for 5
turns after firing for the dynasty who fired him.

Keep in mind, that the Text.dbt from this bugfix mod is german only! So
for other languages, you have to use your own text.dbt and translate it!